New code for the cart page and specific page.
New code in the cartPage to get the product in the cart. New code in the specific page to add and remove product from cart. And the "go to checkout" button takes the user to the checkout page. Otherwise cleaned the App.js file and done some styling in the csss file.