New cart icon code. New "hooks" folder with a new useTheCart file
New code to show a number on the cart icon which shows how many product(s) added to the cart which will display on the checkout page. Major change is that this was done with Zustand, thats why I ceated a new hooks folder with a new useTheCart.jsx file.